Happy Friday!
Just got paged blits working. linear8 runs nicely on my VESA 1.2 laptop
screen now though it's not as fast as the unstable vgagl driver. (paged
framebuffers force doublebuffering on)
Replaced many for loops in linear8 with the inline-asm __mem* functions

Fixed a subtle but possibly disasterous error in handle.c; this could be the
source of the handle tree problems I've been having lately. There is one
special case in htree_delete() in which it must copy the 'identity' of a
handle node from one to another. When I added the 'group' field to the
handle node, I forgot to add a line to copy it. (I added a warning in
handle.h to this effect; hopefully this won't happen again.)
The outcome of all this is that a handle's group can be randomly reassigned
to another handle's group. When one handle is deleted, others might randomly
disappear. This is consistant with observed behavior.

I don't, however, think this is the only problem... Sometimes handles would
appear to swap pointers with each other, or do other strange things. <insert
sound of plane crash>
Unless I find something really obvious I'll make a testbed program I can use
to evaluate the handle functions seperately.

Yep, verified that handles still have problems.

Added a command line switch for double-buffering

Modified svgagl library to support runtime doublebuffering, and removed a
few archaic crufty bits.