Added jpeg support to picogui. I used the jpeg library that's freely available under GPL. (jpeg-6b from the Independent JPEG Group, date 27-Mar-1998. http://www.ijg.org/) It is the most recent release on their download site.
Configuration option is CONFIG_FORMAT_JPEG, and it is disabled
by default. I added it's entry as not defined in profile.defaults.

I have also supported the --with-jpeg-dir=DIR option for the
configure script. This is useful if cross-compiling. DIR needs
to have libjpeg.a (if static), libjpeg.so.* and jpejlib.h in
it, for it to be found correctly. If you do not require a special
version of libjpeg, the configure script should automatically locate
an installed version of the library if it exists.

(If not, go to the website and download and install their latest
version.)

I added a jpeg test file in to the cli_c/demos/data directory.
I tested the new code using imgview in the cli_c/demos directory.
I tested with a SDL/native configuration and a cross-compiled
uclinux version on ARM.