Added server requests for duplicating handles and changing their context.
More importantly though, I think I may have found the source of that
infamous handle bug! In mkhandlep(), the memory is not initialized after
being allocated for the handlenode. Most fields are set immediately
afterwards, but 'group' is not. This could have definitely lead to random
disappearance. If a value from the stack at the time mkhandle is called
happened to match an existing handle, a timebomb would basically be set in
the handle!