This checkin adds a list widget that is mostly functional. It has the side effect of not painting correctly when attached to a scroll widget.
It also adds a menubar widget (which is really a panel that
you can attach widgets to.) The plan for this is to migrate
the ability over to the panel widget and obsolete the menubar
widget. Currently, if you create an application with the
PG_APP_MENUBAR, when you kill the client app, the server will
crash. This won't affect anyone using the standard application
types.

I tested this code using a intel linux pc, using menutest,
helloworld, and another application. I don't anticipate
any of this code affecting current applicatoins. A listtest
app was added to show usage of the list widget.